chore: clean up kube client
