+++ pkg/cmd/seactl/deps/apply.go
@@ -46,6 +46,10 @@ func NewApply() *Apply {
 // RunE generates and applies the dependencies for a development environment.
 func (a *Apply) RunE(cmd *cobra.Command, args []string) error {
 	kubeContext := cmd.Root().Flags().Lookup("context").Value.String()
+	client, err := kube.NewKubectlCmd("", kubeContext)
+	if err != nil {
+		console.Fatal(err.Error())
+	}
 
 	ctx, cancel := signal.NotifyContext(context.Background(), syscall.SIGINT, syscall.SIGTERM)
 	defer cancel()
@@ -63,11 +67,6 @@ func (a *Apply) RunE(cmd *cobra.Command, args []string) error {
 	env, err := manifest.GetEnvironment(args[0])
 	if err != nil {
 		console.Fatal("Build environment '%s' not found in the manifest", args[0])
-	}
-
-	client, err := kube.NewKubectlCmd("", kubeContext)
-	if err != nil {
-		console.Fatal(err.Error())
 	}
 
 	console.Info("Applying dependencies for the '%s' environment", env.Name)
