+++ pkg/kube/client/kubectl_cmd.go
@@ -20,16 +20,19 @@ import (
 	"fmt"
 	"reflect"
 	"regexp"
+	"strings"
 
-	"k8s.io/apimachinery/pkg/api/errors"
+	corev1 "k8s.io/api/core/v1"
+	apierr "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/client-go/discovery"
 	"k8s.io/client-go/dynamic"
 )
 
 // KubectlCmd is the kubernetes client that is used by the seactl tool.
 type KubectlCmd struct {
-	dc     dynamic.Interface
 	client *Client
 }
 
@@ -39,13 +42,7 @@ func NewKubectlCmd(ns, context string) (*KubectlCmd, error) {
 		return nil, err
 	}
 
-	dc, err := c.Factory().DynamicClient()
-	if err != nil {
-		return nil, err
-	}
-
 	return &KubectlCmd{
-		dc:     dc,
 		client: c,
 	}, nil
 }
@@ -56,7 +53,7 @@ func (c *KubectlCmd) Get(ctx context.Context, obj Object, opts metav1.GetOptions
 		return err
 	}
 
-	iface, err := c.client.ResourceInterfaceFor(obj, "get")
+	iface, err := ResourceInterfaceFor(c.client, obj, "get")
 	if err != nil {
 		return err
 	}
@@ -79,7 +76,7 @@ func (c *KubectlCmd) Delete(ctx context.Context, obj Object, opts metav1.DeleteO
 		return err
 	}
 
-	iface, err := c.client.ResourceInterfaceFor(obj, "delete")
+	iface, err := ResourceInterfaceFor(c.client, obj, "delete")
 	if err != nil {
 		return err
 	}
@@ -98,7 +95,7 @@ func (c *KubectlCmd) Create(ctx context.Context, obj Object, opts metav1.CreateO
 		return err
 	}
 
-	iface, err := c.client.ResourceInterfaceFor(obj, "create")
+	iface, err := ResourceInterfaceFor(c.client, obj, "create")
 	if err != nil {
 		return err
 	}
@@ -123,7 +120,7 @@ func (c *KubectlCmd) Update(ctx context.Context, obj Object, opts metav1.UpdateO
 		return err
 	}
 
-	iface, err := c.client.ResourceInterfaceFor(obj, "update")
+	iface, err := ResourceInterfaceFor(c.client, obj, "update")
 	if err != nil {
 		return err
 	}
@@ -143,7 +140,7 @@ func (c *KubectlCmd) Update(ctx context.Context, obj Object, opts metav1.UpdateO
 
 func (c *KubectlCmd) CreateOrUpdate(ctx context.Context, obj Object, f MutateFn) (OperationResult, error) {
 	if err := c.Get(ctx, obj, metav1.GetOptions{}); err != nil {
-		if !errors.IsNotFound(err) {
+		if !apierr.IsNotFound(err) {
 			return OperationResultNone, err
 		}
 
@@ -218,6 +215,61 @@ func PreserveManagedFields(source, target Object) {
 	target.SetAnnotations(annotations)
 }
 
+// ResourceInterfaceFor returns a new dynamic.ResourceInterface for the client.
+func ResourceInterfaceFor(c *Client, obj Object, method string) (dynamic.ResourceInterface, error) {
+	gvk := obj.GetObjectKind().GroupVersionKind()
+
+	dyn, err := c.Factory().DynamicClient()
+	if err != nil {
+		return nil, err
+	}
+
+	dc, err := c.ToDiscoveryClient()
+	if err != nil {
+		return nil, err
+	}
+
+	resource, err := ServerResourcesForGroupVersionKind(dc, gvk, method)
+	if err != nil {
+		return nil, err
+	}
+
+	gvr := gvk.GroupVersion().WithResource(resource.Name)
+	if resource.Namespaced {
+		if obj.GetNamespace() == "" {
+			obj.SetNamespace(corev1.NamespaceDefault)
+		}
+		return dyn.Resource(gvr).Namespace(obj.GetNamespace()), nil
+	}
+	return dyn.Resource(gvr), nil
+}
+
+// ServerResourcesForGroupVersionKind returns the APIResource for the provided GroupVersionKind.
+func ServerResourcesForGroupVersionKind(dc discovery.CachedDiscoveryInterface, gvk schema.GroupVersionKind, verb string) (*metav1.APIResource, error) {
+	resources, err := dc.ServerResourcesForGroupVersion(gvk.GroupVersion().String())
+	if err != nil {
+		return nil, err
+	}
+
+	for _, r := range resources.APIResources {
+		if r.Kind == gvk.Kind {
+			if supportedVerb(&r, verb) {
+				return &r, nil
+			}
+
+			return nil, apierr.NewMethodNotSupported(
+				schema.GroupResource{Group: gvk.Group, Resource: gvk.Kind},
+				verb,
+			)
+		}
+	}
+
+	return nil, apierr.NewNotFound(
+		schema.GroupResource{Group: gvk.Group, Resource: gvk.Kind},
+		"",
+	)
+}
+
 func toUnstructured(obj Object, u *unstructured.Unstructured) error {
 	switch o := obj.(type) { //nolint:gocritic
 	case *unstructured.Unstructured:
@@ -262,4 +314,18 @@ func mutate(f MutateFn, key ObjectKey, obj Object) error {
 		return nil
 	}
 	return nil
+}
+
+// supportedVerb returns true if the provided verb is supported by the APIResource.
+func supportedVerb(apiResource *metav1.APIResource, verb string) bool {
+	if verb == "" || verb == "*" {
+		return true
+	}
+
+	for _, v := range apiResource.Verbs {
+		if strings.EqualFold(v, verb) {
+			return true
+		}
+	}
+	return false
 }
